#131 Make a single timeseries addressable and annotatable

+++ backend/src/main/java/de/dlr/shepard/context/semantic/daos/AnnotatableTimeseriesDAO.java
@@ -12,7 +12,7 @@ import org.neo4j.ogm.cypher.Filter;
 @RequestScoped
 public class AnnotatableTimeseriesDAO extends GenericDAO<AnnotatableTimeseries> {
 
-  public AnnotatableTimeseries findByTimeseries(long containerId, long timeseriesId) {
+  public AnnotatableTimeseries findByTimeseries(long containerId, int timeseriesId) {
     var containerFilter = new Filter("containerId", ComparisonOperator.EQUALS, containerId);
     var timeseriesFilter = new Filter("timeseriesId", ComparisonOperator.EQUALS, timeseriesId);
     return this.session.loadAll(AnnotatableTimeseries.class, containerFilter.and(timeseriesFilter), 2)
